+++ extern/dolphin/include/dolphin/card/CARDFormat.h
@@ -1,6 +1,8 @@
 #ifndef _DOLPHIN_CARDFORMAT_H_
 #define _DOLPHIN_CARDFORMAT_H_
 
-long CARDFormat(long chan);
+#include <dolphin/types.h>
+
+s32 CARDFormat(s32 chan);
 
 #endif // _DOLPHIN_CARDFORMAT_H_
