+++ extern/dolphin/include/dolphin/si.h
@@ -1,6 +1,8 @@
 #ifndef _DOLPHIN_SI_H_
 #define _DOLPHIN_SI_H_
 
+#include <dolphin/types.h>
+
 #define PAD_CHAN0_BIT 0x80000000
 #define PAD_CHAN1_BIT 0x40000000
 #define PAD_CHAN2_BIT 0x20000000
@@ -53,7 +55,7 @@ typedef void (*SITypeAndStatusCallback)(s32 chan, u32 type);
 
 BOOL SITransfer(s32 chan, void *output, u32 outputBytes, void *input, u32 inputBytes,
                 SICallback callback, OSTime delay);
-u32 SIGetCommand(long chan);
+u32 SIGetCommand(s32 chan);
 u32 SIEnablePolling(u32 poll);
 u32 SIDisablePolling(u32 poll);
 u32 SISetXY(u32 x, u32 y);
